# service/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /v2/blackpig/user/by_phone:
        get:
            tags:
                - UserService
            operationId: UserService_FindUserByPhone
            parameters:
                - name: phone
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/blackpig.v2.UserReply'
    /v2/blackpig/user/delete_user:
        post:
            tags:
                - UserService
            operationId: UserService_DeleteUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/blackpig.v2.DeleteUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/blackpig.v2.DeleteUserReply'
    /v2/blackpig/user/list_user:
        get:
            tags:
                - UserService
            operationId: UserService_ListUser
            parameters:
                - name: current
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: size
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: keyWord
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/blackpig.v2.ListReply'
    /v2/blackpig/user/login:
        get:
            tags:
                - UserService
            operationId: UserService_Login
            parameters:
                - name: code
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/blackpig.v2.UserReply'
    /v2/blackpig/user/register:
        post:
            tags:
                - UserService
            operationId: UserService_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/blackpig.v2.User'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/blackpig.v2.RegisterReply'
    /v2/blackpig/user/update_user:
        post:
            tags:
                - UserService
            operationId: UserService_UpdateUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/blackpig.v2.User'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/blackpig.v2.UserReply'
...
components:
    schemas:
...
        blackpig.v2.DeleteUserReply:
            type: object
            properties:
                state:
                    type: string
        blackpig.v2.DeleteUserRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
        blackpig.v2.ListReply:
            type: object
            properties:
                user:
                    type: array
                    items:
                        $ref: '#/components/schemas/blackpig.v2.User'
...
        blackpig.v2.RegisterReply:
            type: object
            properties:
                state:
                    type: string
...
        blackpig.v2.User:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                openid:
                    type: string
                phone:
                    type: string
                username:
                    type: string
                gender:
                    type: string
                carPictures:
                    type: array
                    items:
                        type: string
                description:
                    type: string
        blackpig.v2.UserReply:
            type: object
            properties:
                user:
                    $ref: '#/components/schemas/blackpig.v2.User'
...